[VP-1370,VP-1654,VP-1372]: VCD-CLI support to set/list/remove metadata in vDC Routed Network
VCD-CLI upport to add/update/delete metadata in a vDC Routed Network. Providing key and value adds an entry taking default values of domain (GENERAL), visibility (READWRITE) and metadata value type (MetadataStringValue). Specific valued of domain, visibility and metadata value type can also be provided. API for add and update is same, calling this updates if there is aleady an existing key with this name.

Also fixed Flake-8 formatting errors.

Testing done:
Added test cases for set/list/remove metadata for vDC Routed Network in vdc_network_tests.py. All test of this class passes.
