+++ system_tests/base_config.yaml
@@ -51,6 +51,7 @@ vcd:
   default_ovdc_network_name: 'test-isolated-vdc-network'
   default_ovdc_network_gateway_ip: '10.1.1.1'
   default_ovdc_network_gateway_netmask: '255.255.255.0'
+  default_direct ovdc_network_name: 'test-direct-vdc-network'
 
   default_catalog_name: 'test-cat'
   default_template_file_name: 'test_vapp_template.ova'
@@ -89,11 +90,11 @@ pvdc:
   resource_pool_names: ['resource-pool-name-1', 'resource-pool-name-2', ...]
 
 external_network:
-  name: '<ext-net-name>'
+  name: 'ExtNw'
   port_group_name: '<pg-name>'
   ip_scopes:
-    - subnet: '<gateway-ip/prefix_len>'
+    - subnet: '2.2.3.1/20'
       ip_ranges:
-        - '<start-ip>-<end-ip>'
-  gateway_sub_allocated_ip_range: '<ip-range>'
-  new_gateway_sub_allocated_ip_range: '<ip-range>'
+        - '2.2.3.2-2.2.3.100'
+  gateway_sub_allocated_ip_range: '2.2.3.5-2.2.3.10'
+  new_gateway_sub_allocated_ip_range: '2.2.3.11-2.2.3.12'
